+++ src/main/java/com/verdantartifice/thaumicwonders/common/misc/GuiIds.java
@@ -5,4 +5,5 @@ public class GuiIds {
     public static final int CATALYZATION_CHAMBER = 2;
     public static final int METEORB = 3;
     public static final int MEATY_ORB = 4;
+    public static final int STRUCTURE_DIVINER = 5;
 }